Split out blaming from extension
