+++ test/extension.test.ts
@@ -1,13 +1,15 @@
 import * as assert from "assert";
 
+import { GitExtension } from "../src/git/extension";
 import { GitBlame } from "../src/git/blame";
 
 suite("Web URL formatting", (): void => {
     const blame = new GitBlame();
+    const extension = new GitExtension(blame);
 
     test("https://", (): void => {
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "https://example.com/user/repo.git",
                 "hash",
                 false,
@@ -15,7 +17,7 @@ suite("Web URL formatting", (): void => {
             "https://example.com/user/repo/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "https://example.com/user/repo",
                 "hash",
                 false,
@@ -26,7 +28,7 @@ suite("Web URL formatting", (): void => {
 
     test("git@", (): void => {
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "git@example.com:user/repo.git",
                 "hash",
                 false,
@@ -34,14 +36,18 @@ suite("Web URL formatting", (): void => {
             "https://example.com/user/repo/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath("git@example.com:user/repo", "hash", false),
+            extension.defaultWebPath(
+                "git@example.com:user/repo",
+                "hash",
+                false,
+            ),
             "https://example.com/user/repo/commit/hash",
         );
     });
 
     test("username@", (): void => {
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "username@example.com:user/repo.git",
                 "hash",
                 false,
@@ -49,7 +55,7 @@ suite("Web URL formatting", (): void => {
             "https://example.com/user/repo/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "username@example.com:user/repo",
                 "hash",
                 false,
@@ -60,7 +66,7 @@ suite("Web URL formatting", (): void => {
 
     test("https:// with port", (): void => {
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "https://example.com:8080/user/repo.git",
                 "hash",
                 false,
@@ -68,7 +74,7 @@ suite("Web URL formatting", (): void => {
             "https://example.com/user/repo/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "https://example.com:8080/user/repo",
                 "hash",
                 false,
@@ -79,7 +85,7 @@ suite("Web URL formatting", (): void => {
 
     test("git@ with port", (): void => {
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "git@example.com:8080/user/repo.git",
                 "hash",
                 false,
@@ -87,7 +93,7 @@ suite("Web URL formatting", (): void => {
             "https://example.com/user/repo/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "git@example.com:8080/user/repo",
                 "hash",
                 false,
@@ -98,7 +104,7 @@ suite("Web URL formatting", (): void => {
 
     test("https:// plural", (): void => {
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "https://example.com/user/repo.git",
                 "hash",
                 true,
@@ -106,14 +112,18 @@ suite("Web URL formatting", (): void => {
             "https://example.com/user/repo/commits/hash",
         );
         assert.equal(
-            blame.defaultWebPath("https://example.com/user/repo", "hash", true),
+            extension.defaultWebPath(
+                "https://example.com/user/repo",
+                "hash",
+                true,
+            ),
             "https://example.com/user/repo/commits/hash",
         );
     });
 
     test("ssh:// short host no user", (): void => {
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "ssh://user@host:8080/SomeProject.git",
                 "hash",
                 false,
@@ -121,7 +131,7 @@ suite("Web URL formatting", (): void => {
             "https://host/SomeProject/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "ssh://user@host:8080/SomeProject",
                 "hash",
                 false,
@@ -132,7 +142,7 @@ suite("Web URL formatting", (): void => {
 
     test("non-alphanumeric in path", (): void => {
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "https://example.com/us.er/repo.git",
                 "hash",
                 false,
@@ -140,7 +150,7 @@ suite("Web URL formatting", (): void => {
             "https://example.com/us.er/repo/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "https://example.com/user/re-po.git",
                 "hash",
                 false,
@@ -148,7 +158,7 @@ suite("Web URL formatting", (): void => {
             "https://example.com/user/re-po/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "https://example.com/user/re%20po.git",
                 "hash",
                 false,
@@ -156,7 +166,7 @@ suite("Web URL formatting", (): void => {
             "https://example.com/user/re%20po/commit/hash",
         );
         assert.equal(
-            blame.defaultWebPath(
+            extension.defaultWebPath(
                 "ssh://user@example.com:us.er/repo.git",
                 "hash",
                 false,
@@ -169,38 +179,41 @@ suite("Web URL formatting", (): void => {
 
 suite("Origin to project name", (): void => {
     const blame = new GitBlame();
+    const extension = new GitExtension(blame);
 
     test("https://", (): void => {
         assert.equal(
-            blame.projectNameFromOrigin("https://example.com/user/repo.git"),
+            extension.projectNameFromOrigin(
+                "https://example.com/user/repo.git",
+            ),
             "repo",
         );
         assert.equal(
-            blame.projectNameFromOrigin("https://example.com/user/repo"),
+            extension.projectNameFromOrigin("https://example.com/user/repo"),
             "repo",
         );
     });
 
     test("git@", (): void => {
         assert.equal(
-            blame.projectNameFromOrigin("git@example.com/user/repo.git"),
+            extension.projectNameFromOrigin("git@example.com/user/repo.git"),
             "repo",
         );
         assert.equal(
-            blame.projectNameFromOrigin("git@example.com/user/repo"),
+            extension.projectNameFromOrigin("git@example.com/user/repo"),
             "repo",
         );
     });
 
     test("longer than normal path", (): void => {
         assert.equal(
-            blame.projectNameFromOrigin(
+            extension.projectNameFromOrigin(
                 "git@example.com/company/group/user/repo.git",
             ),
             "repo",
         );
         assert.equal(
-            blame.projectNameFromOrigin(
+            extension.projectNameFromOrigin(
                 "git@example.com/company/group/user/repo",
             ),
             "repo",
@@ -209,25 +222,25 @@ suite("Origin to project name", (): void => {
 
     test("non-alphanumeric in path", (): void => {
         assert.equal(
-            blame.projectNameFromOrigin(
+            extension.projectNameFromOrigin(
                 "https://example.com/user/re-po.git",
             ),
             "re-po",
         );
         assert.equal(
-            blame.projectNameFromOrigin(
+            extension.projectNameFromOrigin(
                 "https://example.com/us.er/repo.git",
             ),
             "repo",
         );
         assert.equal(
-            blame.projectNameFromOrigin(
+            extension.projectNameFromOrigin(
                 "https://example.com/user/re.po.git",
             ),
             "re.po",
         );
         assert.equal(
-            blame.projectNameFromOrigin(
+            extension.projectNameFromOrigin(
                 "https://example.com/user/re.po",
             ),
             "re.po",
