+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
     "compilerOptions": {
         "module": "commonjs",
-        "target": "es6",
+        "target": "es2018",
         "strict": true,
-        "outDir": "out",
+        "outDir": "./out",
         "lib": [
             "es2016",
             "es2017",
